Rename files to better indicates that these constants are fake test secrets
