+++ sdk/storage/storage-blob/test/blobclient.spec.ts
@@ -24,7 +24,7 @@ import {
   BlobServiceClient,
   RehydratePriority
 } from "../src";
-import { Test_CPK_INFO } from "./utils/constants";
+import { Test_CPK_INFO } from "./utils/fakeTestSecrets";
 import { base64encode } from "../src/utils/utils.common";
 import { context, setSpan } from "@azure/core-tracing";
 import { Context } from "mocha";
