+++ sdk/storage/storage-blob/test/blockblobclient.spec.ts
@@ -13,7 +13,7 @@ import {
   recorderEnvSetup
 } from "./utils";
 import { ContainerClient, BlobClient, BlockBlobClient } from "../src";
-import { Test_CPK_INFO } from "./utils/constants";
+import { Test_CPK_INFO } from "./utils/fakeTestSecrets";
 import { BlockBlobTier } from "../src";
 import { Context } from "mocha";
 import { isNode } from "@azure/core-http";
