+++ sdk/storage/storage-blob/test/encrytion.spec.ts
@@ -6,7 +6,7 @@ import * as dotenv from "dotenv";
 import { getBSU, recorderEnvSetup } from "./utils";
 import { record, isPlaybackMode, Recorder } from "@azure-tools/test-recorder";
 import { BlobServiceClient, BlobClient, BlockBlobClient, ContainerClient } from "../src";
-import { Test_CPK_INFO } from "./utils/constants";
+import { Test_CPK_INFO } from "./utils/fakeTestSecrets";
 import { isNode } from "@azure/core-http";
 import { Context } from "mocha";
 dotenv.config();
