+++ sdk/storage/storage-blob/test/node/highlevel.node.spec.ts
@@ -20,7 +20,7 @@ import { record, Recorder } from "@azure-tools/test-recorder";
 import { ContainerClient, BlobClient, BlockBlobClient, BlobServiceClient } from "../../src";
 import { readStreamToLocalFileWithLogs } from "../utils/testutils.node";
 import { BLOCK_BLOB_MAX_STAGE_BLOCK_BYTES } from "../../src/utils/constants";
-import { Test_CPK_INFO } from "../utils/constants";
+import { Test_CPK_INFO } from "../utils/fakeTestSecrets";
 import { streamToBuffer2 } from "../../src/utils/utils.node";
 import { delay } from "../../src/utils/utils.common";
 import { Context } from "mocha";
